[Develop 5주차] 생명석 시스템/총 내려놓기 시스템 구현

+++ Source/QuadLand/UI/QLPlayerHpBarWidget.h
@@ -28,7 +28,7 @@ protected:
 	UPROPERTY()
 	TObjectPtr<class UProgressBar> HP;
 
-	int CurrentHP;
+	float CurrentHP;
 
-	int MaxHP;
+	float MaxHP;
 };
